feat: restructure HomePage component and move LandingPage.
